improve error messages and password change on email confirmation. fix #146
