--- dashboard/src/app/modules/auth/components/form/forgottenPassword/template.html
@@ -1,15 +1,0 @@
-<form (ngSubmit)="onSubmit()">
-  <div class="p-grid ui-fluid" [formGroup]="forgottenPasswordForm">
-    <div class="p-col-12">
-      <input type="email" pInputText formControlName="email" placeholder="Email">
-      <p-message
-          severity="error"
-          [text]="forgottenPasswordForm.controls['email'].errors['pattern'] ? 'L\'email est invalide.' : ''"
-          *ngIf="!forgottenPasswordForm.controls['email'].valid&&submitted&&forgottenPasswordForm.controls['email'].value">
-      </p-message>
-    </div>
-  </div>
-  <p>
-    <p-button (onClick)="onSubmit" label="Envoyer" [disabled]="!forgottenPasswordForm.controls['email'].valid"></p-button>
-  </p>
-</form>
