+++ dashboard/src/app/modules/auth/pages/confirm-email/confirm-email.page.ts
@@ -3,10 +3,13 @@ import { ActivatedRoute, Params } from '@angular/router';
 import { Message } from 'primeng/api';
 
 import { AuthenticationService } from '~/applicativeService/authentication/auth.service';
+import { MAIN } from '~/config/main';
 
 @Component({ templateUrl: 'template.html' })
 export class AuthPageConfirmEmailComponent implements OnInit {
   public messages: Message[] = [];
+  public hasError = false;
+  public contactEmail: string = MAIN.contactEmail;
 
   constructor(private activatedRoute: ActivatedRoute, private authenticationService: AuthenticationService) {}
 
@@ -17,22 +20,22 @@ export class AuthPageConfirmEmailComponent implements OnInit {
 
       this.authenticationService.checkEmailToken(reset, token).subscribe(
         () => {
+          this.hasError = false;
           this.messages = [
             {
               severity: 'success',
               summary: 'Email confirmé',
-              detail: `Vous pouvez maintenant vous connecter.
-                       <br><br>
-                       <a href="/signin">Se connecter</a>`,
+              detail: 'Vous pouvez changer votre mot de passe',
             },
           ];
         },
         (error) => {
+          this.hasError = true;
           this.messages = [
             {
               severity: 'error',
-              summary: error.statusText,
-              detail: error.message,
+              summary: 'Email non confirmé',
+              detail: null,
             },
           ];
         },
