+++ dashboard/src/app/modules/auth/pages/confirm-email/template.html
@@ -1,0 +1,14 @@
+<section class="auth__wrapper">
+  <div class="has-text-centered">
+    <h3>Confirmation de votre adresse email.</h3>
+    <!-- display a password change field if the email is verified -->
+    <app-auth-form-new-password *ngIf="!hasError"></app-auth-form-new-password>
+
+    <p-messages [value]="messages" [closable]="false"></p-messages>
+
+    <p *ngIf="hasError">
+      Merci de contacter <a href="mailto:{{ contactEmail }}">{{ contactEmail }}</a>
+      pour faire une demande d'accès au service.
+    </p>
+  </div>
+</section>
