--- dashboard/src/app/modules/auth/pages/confirmEmail/template.html
@@ -1,8 +1,0 @@
-<section class="auth__wrapper">
-  <div class="has-text-centered">
-    <h3>Confirmation de votre adresse email.</h3>
-
-    <p-messages [value]="messages" [closable]="false"></p-messages>
-
-  </div>
-</section>
