--- dashboard/src/app/modules/auth/pages/newPassword/component.ts
@@ -1,7 +1,0 @@
-import { Component } from '@angular/core';
-
-@Component({
-  templateUrl: 'template.html',
-  styleUrls: ['style.scss'],
-})
-export class AuthPageNewPasswordComponent {}
