--- dashboard/src/app/modules/auth/pages/newPassword/template.html
@@ -1,7 +1,0 @@
-<section class="auth__wrapper">
-  <div class="has-text-centered">
-    <h3>Changement de mot de passe</h3>
-    <p>Veuillez renseigner votre nouveau mot de passe</p>
-    <app-auth-form-new-password></app-auth-form-new-password>
-  </div>
-</section>
