+++ dashboard/src/app/modules/auth/pages/signin/style.scss
@@ -1,4 +1,4 @@
-.box{
+.box {
   display: flex;
   flex-wrap: wrap;
   flex-direction: column;
