+++ dashboard/src/app/modules/auth/pages/signin/template.html
@@ -1,14 +1,14 @@
 <section class="auth__wrapper">
   <div class="has-text-centered">
-    <img src="assets/icons/car.svg" alt="logo" width="200">
+    <img src="assets/icons/car.svg" alt="logo" width="200" />
 
     <h3>Registre de Preuve de Covoiturage</h3>
 
     <app-auth-form-signin></app-auth-form-signin>
 
     <p>
-      <a routerLink="/forgotten-password">Mot de passe oublié&nbsp;?</a><br>
-      <a href="mailto:{{contactEmail}}">Demander la création d'un compte</a><br>
+      <a routerLink="/forgotten-password">Mot de passe oublié&nbsp;?</a><br />
+      <a href="mailto:{{ contactEmail }}">Demander la création d'un compte</a><br />
       <a href="http://covoiturage.beta.gouv.fr">Retour sur covoiturage.beta.gouv.fr</a>
     </p>
   </div>
