+++ dashboard/tslint.json
@@ -23,7 +23,7 @@
     "no-console": [true, "debug", "info", "time", "timeEnd", "trace"],
     "no-debugger": true,
     "no-duplicate-super": true,
-    "no-empty": true,
+    "no-empty": false,
     "no-empty-interface": true,
     "no-inferrable-types": [true, "ignore-params"],
     "no-misused-new": true,
